+++ src/components/grid/ColRow.tsx
@@ -1,15 +1,14 @@
+/* tslint:disable */
 import {IStyle} from 'fela'
 import * as React from 'react'
 import {FelaComponent, StyleFunction} from 'react-fela'
 import {BaseTheme} from '../../baseTheme'
 import {createPaddingCSSProps, createStylingCSSProps, mergeThemedStyles, PaddingProps, StylingProps, WithStyle} from '../base'
-
+import {GRID_SIZE} from './Col'
 import {GridContext} from './GridContext'
 import {testGridContext} from './testGridContext'
 
 interface Props<T extends BaseTheme> extends WithStyle<T>, PaddingProps, StylingProps {
-  height?: number | string,
-  minHeight?: number | string,
 }
 
 interface RowContainerProps<T extends BaseTheme> extends Props<T> {
@@ -18,30 +17,21 @@ interface RowContainerProps<T extends BaseTheme> extends Props<T> {
 
 interface RowStyle extends IStyle {
   ':first-child': IStyle,
-  '@media (min-width: 768px)': IStyle,
 }
 
 class RowContainer<T extends BaseTheme> extends React.Component<RowContainerProps<T>> {
 
   public render() {
-    const rowStyle: StyleFunction<BaseTheme, RowContainerProps<T>> = ({style, minHeight, spacing, height, ...otherProps}): RowStyle => ({
+    const rowStyle: StyleFunction<BaseTheme, RowContainerProps<T>> = ({style, spacing, ...otherProps}): RowStyle => ({
       ...createPaddingCSSProps(otherProps),
       ...createStylingCSSProps(otherProps),
       boxSizing: 'border-box',
       display: 'flex',
-      // wrap all flex items -> since a panel has a mobile width of 100%, each
-      // panel visually gets its own row
       flexWrap: 'wrap',
-      // let all content items (= panels) claim the full space
-      alignItems: 'stretch',
       width: '100%',
-      minHeight,
+      marginTop: spacing,
       ':first-child': {
-        marginTop: `-${spacing}`,
-      },
-      '@media (min-width: 768px)': {
-        flexWrap: 'nowrap',
-        height,
+        marginTop: 0,
       },
     })
     const {children, style, ...otherProps} = this.props
@@ -54,11 +44,41 @@ class RowContainer<T extends BaseTheme> extends React.Component<RowContainerProp
   }
 }
 
-export class Row<T extends BaseTheme> extends React.Component<Props<T>> {
+export class ColRow<T extends BaseTheme> extends React.Component<Props<T>> {
 
-  public static defaultProps = {
-    height: 'auto',
-    minHeight: 'auto',
+  renderChildren(spacing: string | number) {
+    let currentRowSize = 0
+    let rowsCount = 0
+
+    // see https://mxstbr.blog/2017/02/react-children-deepdive/#looping-over-children
+    return React.Children.map(this.props.children, (child) => {
+      const currentChild = child as any
+      // calculate the sum of all <Col> sizes
+      currentRowSize += currentChild.props.size
+
+      // the <Col> child fills up the full space -> manipulate rowBreak and marginTop
+      if (currentRowSize === GRID_SIZE) {
+        currentRowSize = 0
+        rowsCount++
+        return React.cloneElement((currentChild), {
+          rowBreak: true,
+          marginTop: rowsCount > 1 ? spacing : 0
+        })
+      }
+
+      // increase the row count if the current <Col> will be rendered at the next line
+      if (currentRowSize >= GRID_SIZE) {
+        rowsCount++
+      }
+
+      // for all rows except the first manipulate the marginTop of the <Col> child
+      if (rowsCount > 0) {
+        return React.cloneElement((currentChild), {
+          marginTop: spacing
+        })
+      }
+      return currentChild
+    })
   }
 
   public render() {
@@ -67,7 +87,7 @@ export class Row<T extends BaseTheme> extends React.Component<Props<T>> {
         {({spacing}) => {
           const child = (
             <RowContainer spacing={spacing} {...this.props}>
-              {this.props.children}
+              {this.renderChildren(spacing)}
             </RowContainer>
           )
           return testGridContext(spacing, child)