+++ src/components/grid/GridContext.ts
@@ -1,7 +1,9 @@
 import * as React from 'react'
 
+export type Spacing = number | string
+
 type Context = {
-  spacing: number | string,
+  spacing: Spacing,
 }
 
 export const GridContext = React.createContext<Context>({